requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
djangorestframework==3.14.0
idna==3.4
Pillow==9.3.0
pytz==2022.7.1
requests==2.28.1
sqlparse==0.4.3
stripe==4.2.0
urllib3==1.26.12
